+++ app/controllers/concerns/authentications.rb
@@ -1,25 +1,25 @@
-# module Authentications
-#   extend ActiveSupport::Concern
+module Authentications
+  extend ActiveSupport::Concern
 
-#   def log_in(user)
-#     session[:user_id] = user.id
-#   end
+  def log_in(user)
+    session[:user_id] = user.id
+  end
 
-#   def current_user
-#     @current_user ||= User.find_by(id: session[:user_id])
-#   end
+  def current_user
+    @current_user ||= User.find_by(id: session[:user_id])
+  end
 
-#   def logged_in?
-#     !current_user.nil?
-#   end
+  def logged_in?
+    !current_user.nil?
+  end
 
-#   def log_out
-#     session.delete(:user_id)
-#     @current_user = nil
-#   end
+  def log_out
+    session.delete(:user_id)
+    @current_user = nil
+  end
 
-#   def authorize
-#     redirect_to '/login' unless current_user
-#   end
+  def authorize
+    redirect_to '/login' unless current_user
+  end
 
-# end
+end
