+++ app/controllers/messages_controller.rb
@@ -1,6 +1,6 @@
 # module V1
   class MessagesController < ApplicationController
-    skip_before_action :authenticate_user_from_token!, only: [:index, :create]
+
 
     def index
       @house = House.find(params[:house_id])
@@ -11,10 +11,10 @@
 
   def create
     @house = House.find(params[:house_id])
-    @messages = Message.where(house_id: @house.id)
+    @user = current_user
 
     @message = @house.messages.new(message_params)
-    @message.update_attributes(author: current_user)
+    @message.update_attributes(author: @user.first_name)
 
     if @message.save
       render json: @message, each_serializer: MessagesSerializer
