+++ app/models/user.rb
@@ -8,7 +8,6 @@ class User < ActiveRecord::Base
   has_attached_file :avatar, :styles => { :medium => "300x300>", :thumb => "100x100>" }, :default_url => "/images/:style/missing.png"
   validates_attachment_content_type :avatar, :content_type => /\Aimage\/.*\Z/
 
-  has_many :messages
   has_many :housing_assignments
   has_many :houses, through: :housing_assignments
   has_many :events
