+++ app/serializers/messages_serializer.rb
@@ -1,7 +1,6 @@
 class MessagesSerializer < ActiveModel::Serializer
-  embed :ids, :include => true
+  # embed :ids, :include => true
 
-  attributes :content, :created_at, :id
-  has_one :author, class_name: "User"
-  has_one :house
+  attributes :content, :created_at, :id, :author
+  # has_one :house
 end
