+++ db/migrate/20150416212824_create_messages.rb
@@ -2,7 +2,7 @@ class CreateMessages < ActiveRecord::Migration
   def change
     create_table :messages do |t|
       t.string :content
-      t.belongs_to :user
+      t.string :author
       t.references :house, index: true, foreign_key: true
 
       t.timestamps null: false
