+++ db/seeds.rb
@@ -10,7 +10,7 @@ chore = Chore.create(task: "Do the dishes", house_id: 1)
 user.chores.create(task: "Mop the floor", house_id: 1)
 event = user.events.create(name: "Halloween Party", date: "2013-09-12 22:49:27", description: "Costume party!!", house_id: 1)
 communal_item = CommunalItem.create(name: "toilet paper", brand: "charmin", quantity: 12, house_id: 1, stock_level: "low")
-message = user.messages.create(content: "You're the best!!", house_id: 1)
+message = Message.create(content: "You're the best!!", house_id: 1, author: "Paul")
 chore_issue = chore.issues.create(reason: "whatever", user_id: 1)
 item_issue = communal_item.issues.create(reason: "I hate paper towel", user_id: 1)
 event_issue = event.issues.create(reason: "I hate halloween", user_id: 1)
@@ -25,5 +25,5 @@ chore3 = Chore.create(task: "Wash your face", house_id: 1)
 chore4 = Chore.create(task: "Get out of my face", house_id: 1)
 chore5 = Chore.create(task: "Say hello like a fucking normal person", house_id: 1)
 
-message2 = user.messages.create(content: "Love you!!!", house_id: 1)
-message3 = user.messages.create(content: "OMFG, You guys suck!", house_id: 1)
+message2 = Message.create(content: "Love you!!!", house_id: 1, author: "Paul")
+message3 = Message.create(content: "OMFG, You guys suck!", house_id: 1, author: "Paul")
